requirements.txt
fastapi
uvicorn
python-multipart
pydantic
torch
torchvision
pillow
numpy
python-dotenv
requests
beautifulsoup4
ultralytics
